Split out header component and remove theme-ui layout components
